Fixes issues with human Life() processing.

+++ code/modules/mob/living/simple_animal/friendly/farm_animals.dm
@@ -67,7 +67,9 @@
 
 /mob/living/simple_animal/hostile/retaliate/goat/handle_living_non_stasis_processes()
 	. = ..()
-	if(. && stat == CONSCIOUS && udder && prob(5))
+	if(!.)
+		return FALSE
+	if(stat == CONSCIOUS && udder && prob(5))
 		udder.add_reagent(/decl/material/liquid/drink/milk, rand(5, 10))
 
 /mob/living/simple_animal/hostile/retaliate/goat/Retaliate()
@@ -141,7 +143,9 @@
 
 /mob/living/simple_animal/cow/handle_living_non_stasis_processes()
 	. = ..()
-	if(. && udder && prob(5))
+	if(!.)
+		return FALSE
+	if(udder && prob(5))
 		udder.add_reagent(/decl/material/liquid/drink/milk, rand(5, 10))
 
 /mob/living/simple_animal/cow/default_disarm_interaction(mob/user)
@@ -185,11 +189,12 @@
 
 /mob/living/simple_animal/chick/handle_living_non_stasis_processes()
 	. = ..()
-	if(.)
-		amount_grown += rand(1,2)
-		if(amount_grown >= 100)
-			new /mob/living/simple_animal/chicken(src.loc)
-			qdel(src)
+	if(!.)
+		return FALSE
+	amount_grown += rand(1,2)
+	if(amount_grown >= 100)
+		new /mob/living/simple_animal/chicken(src.loc)
+		qdel(src)
 
 var/global/const/MAX_CHICKENS = 50
 var/global/chicken_count = 0
@@ -251,7 +256,9 @@ var/global/chicken_count = 0
 
 /mob/living/simple_animal/chicken/handle_living_non_stasis_processes()
 	. = ..()
-	if(. && prob(3) && eggsleft > 0)
+	if(!.)
+		return FALSE
+	if(prob(3) && eggsleft > 0)
 		visible_message("[src] [pick("lays an egg.","squats down and croons.","begins making a huge racket.","begins clucking raucously.")]")
 		eggsleft--
 		var/obj/item/chems/food/egg/E = new(get_turf(src))
